fix(storage): Fix save password filed length as int64
Fix save password filed length as int64:
* Correct length type from int64 to uint32.

+++ keycontainerex-backend/storage/storage_save.go
@@ -69,7 +69,7 @@ func (s *Storage) makePasswordSection() []byte {
 			default:
 				continue
 			}
-			_ = binary.Write(buffer, binary.LittleEndian, int64(length))
+			_ = binary.Write(buffer, binary.LittleEndian, uint32(length))
 			buffer.Write(fieldValueBytes)
 		}
 		buffer.WriteByte(passwordSplit)
